Client can now Search with Results!!!!
PPAS-3638 #comment Client can now Search with Results!!!! Both the client and the server-side code needs additional work to iron out the kinks - e.g. the client can't currently create a report because {{commons-logging}} has delibarately been excluded from the classpath, and JasperReports requires it... next to investigate how to properly wire it up with the rest of the logging framework.