PPAS returns Strings in nested composite types
PPAS-3680 #comment Add generic unit/integration testing database for the JDBC project - now it can stand completely alone! Yay!
PPAS-3680 #comment PPAS returns Strings in nested composite types - Updated JDBC to support parsing both String => java.sql.Array and java.sql.Struct
PPAS-3680 #comment In order to manage the coercion of String to JDBC type, the Adapter abstraction had to leak into the model machinery - prior to this commit, the Struct and Array classes had no dependencies on either the Oracle or the EDB implementations of the JDBC API, but now they (and the model to a small degree) are required to understand something of the EDB API. THIS IS A BAD THING, but unavoidable.