Make Metrics client Deployment and Cluster Node "aware"
PPAS-3843 #comment Completely new Run/Debug configurations. The IntelliJ JBoss integration is broken for domain mode, therefore a set of Maven run goals and Remote configurations were created to reproduce that integration: 1) Wildfly Start-up and Shutdown, 2) Wildfly servers deployment.
PPAS-3843 #comment There are currently two servers groups, two profiles and two servers: oracle and edb, where the datasources for each database is associated with the appropriate profile, which in turn is assigned to the appropriate server group, and thus the nodes running in their respective groups can use the same codebase, with the same datasource names, but targeting different databases.
PPAS-3843 #comment By adding a new column and attribute to the Metrics User - node_target, which must match the available server names provisioned in Wildfly. Those names are currently 'oracle1' and 'edb1'. The client is able to filter available nodes and pick from those that match the node_target attribute.
PPAS-3843 #comment If the user has never logged in, the client will use the Oracle server for a single query to determine which node_target is appropriate, and then direct all further requests to the node_target selected nodes.
PPAS-3843 #comment If the user has logged in, but the node_target has not been set, the client will use the Oracle server for a single query to determine which node_target is appropriate, and then direct all further requests to the node_target selected nodes.
PPAS-3843 #comment If the user has logged in, and their saved node_target value is different than the value from the database, use the value from the database.
PPAS-3843 #comment Included utility script for adding multicast routes on Mac - apparently this isn't a default, and must be done manually. Use the resource-filtered "add-multicast.sh", as it will be automatically filled with your local IP address, which is mandatory.