Fixes in 9.5.1.6 leave EMPTY arrays as strings
Version 9.5.1.6 of EDB JDBC driver leaves EMPTY arrays as strings -
this means the string processing needs to remain.  Unfortunately, the
string parsing function that has been placed in the driver now requires
a Connection object in order to parse Structs - this is less than ideal for
our setup, as the JdbcArrayList and StructArrayList currently don't have
a reference to the JDBC connection - nor should they need one.

This has forced another pull of code directly from the driver into our
codebase in order to parse strings as Structs.